feat: add feature submodules
