+++ botcopier/features/__init__.py
@@ -1,0 +1,14 @@
+"""Convenience exports for feature utilities."""
+from .anomaly import _clip_apply, _clip_train_features, _score_anomalies
+from .augmentation import _augment_dataframe, _augment_dtw_dataframe
+from .technical import _extract_features, _neutralize_against_market_index
+
+__all__ = [
+    "_augment_dataframe",
+    "_augment_dtw_dataframe",
+    "_clip_train_features",
+    "_clip_apply",
+    "_score_anomalies",
+    "_extract_features",
+    "_neutralize_against_market_index",
+]
